requirements.txt
fastapi
uvicorn
psycopg2
pandas
requests
sentence-transformers
faiss-cpu
matplotlib
streamlit
